requirements.txt
cycler==0.10.0
dill==0.2.5
dj-database-url==0.4.2
Django==2.0
django-estimators==0.2.1
matplotlib==2.1.0
numpy==1.13.3
pandas==0.21.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
whitenoise==3.3.1
